refactor - moved all condition related data from battle component to condition component
