+++ src/app/models/actorsData.ts
@@ -1,8 +1,8 @@
 import {Actor} from "./actor";
 
 export const PROTAGONISTS = [
-  new Actor('Belmara', 35, 35, 4),
-  new Actor('Kudo', 32, 32, 3),
-  new Actor('Meaturus', 14, 14, 2),
-  new Actor('Sir Mikhail', 25, 25, 1),
+  new Actor('Belmara', 53, 53, 4),
+  new Actor('Kudo', 49, 49, 3),
+  new Actor('Meaturus', 31, 31, 2),
+  new Actor('Sir Mikhail', 36, 36, 1),
 ];
